Detach support

+++ hume/device_controller/device_controller/device/device_procedures.py
@@ -46,12 +46,13 @@ def confirm_attach(device_uuid):
     :param device_uuid: device to attach
     :type device_uuid: str
     """
-    LOGGER.debug(f"confirm attach for: {device_uuid}")
-
     device = hume_storage.get(Device, device_uuid)
 
     if device:
+
         if device_req_lib.heartbeat_request(device):
+            LOGGER.debug("heartbeat request succeeded")
+
             device.attached = True
             hume_storage.save(device)
 
@@ -64,8 +65,10 @@ def confirm_attach(device_uuid):
             return
         else:
             LOGGER.warning("device did not respond to heartbeat request")
-            # Remove, either the device info is faulty, or device has issues
-            hume_storage.delete(device)
+            # Remove if not attached, either the device info is faulty, or
+            # device has issues
+            if not device.attached:
+                hume_storage.delete(device)
     else:
         LOGGER.error("device to be attached does not exist")
 
@@ -75,3 +78,23 @@ def confirm_attach(device_uuid):
             "content": {"device_uuid": device_uuid, "success": False}
         }
     )
+
+
+def detach(device_uuid):
+    """
+    Detach the parameter device from HUME, removing all associated data.
+
+    NOTE! In the future, this procedure needs to be extended to delete
+    configurations as well.
+
+    :param device_uuid: UUID of device to detach
+    :type device_uuid: str
+    """
+    device = hume_storage.get(Device, device_uuid)
+
+    if device:
+        LOGGER.debug("found device to delete")
+
+        hume_storage.delete(device)
+    else:
+        LOGGER.debug("device does not exist")
